LayoutTests/svg/custom/use-event-listener-passive-option-preserved-expected.txt

Tests that the 'passive' option on an event listener is preserved when copied to an SVG use shadow tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS defaultPrevented is false
PASS successfullyParsed is true

TEST COMPLETE
